# express-back/docs/swagger.generated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "哈尔滨学院校园综合服务平台 API",
    "version": "1.0.0",
    "description": "校园综合服务平台后端接口文档。"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Current server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/admin/services": {
      "get": {
        "tags": [
          "Admin Services"
        ],
        "summary": "Get Services",
        "operationId": "AdminServiceController_getServices",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Admin Services"
        ],
        "summary": "Create Service",
        "operationId": "AdminServiceController_createService",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/services/{id}": {
      "put": {
        "tags": [
          "Admin Services"
        ],
        "summary": "Update Service",
        "operationId": "AdminServiceController_updateService",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/services/{id}/status": {
      "patch": {
        "tags": [
          "Admin Services"
        ],
        "summary": "Update Service Status",
        "operationId": "AdminServiceController_updateServiceStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/service/auth/login": {
      "post": {
        "tags": [
          "Service Auth"
        ],
        "summary": "客服登录",
        "operationId": "ServiceAuthController_login",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true
      }
    },
    "/api/service/auth/profile": {
      "get": {
        "tags": [
          "Service Auth"
        ],
        "summary": "需要认证的客服路由",
        "operationId": "ServiceAuthController_getProfile",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Service Auth"
        ],
        "summary": "Update Profile",
        "operationId": "ServiceAuthController_updateProfile",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/service/auth/change-password": {
      "post": {
        "tags": [
          "Service Auth"
        ],
        "summary": "Change Password",
        "operationId": "ServiceAuthController_changePassword",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/service/auth/logout": {
      "post": {
        "tags": [
          "Service Auth"
        ],
        "summary": "Logout",
        "operationId": "ServiceAuthController_logout",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}